+++ Packages/com.mimylab.fukuroudon/CHANGELOG.md
@@ -6,6 +6,12 @@
 
 利用手順は[こちら](https://github.com/mimyquality/FukuroUdon/wiki)からご確認ください。
 
+## [1.18.0] - 2024/4/16
+
+- **Added**
+  - Active Relay
+    - Active Relay to GameObjectを追加しました。
+
 ## [1.17.0] - 2024/4/15
 
 - **Added**
@@ -417,6 +423,7 @@
     - 0.2.7バージョンで、later-joiner視点でjoin直後が非アクティブだと同期しなかったのを修正しました。
     - 他人がピックアップした際にオブジェクトが一瞬跳ねる現象を低減しました。
 
+[1.18.0]: https://github.com/mimyquality/FukuroUdon/releases/tag/1.18.0
 [1.17.0]: https://github.com/mimyquality/FukuroUdon/releases/tag/1.17.0
 [1.16.4]: https://github.com/mimyquality/FukuroUdon/releases/tag/1.16.4
 [1.16.3]: https://github.com/mimyquality/FukuroUdon/releases/tag/1.16.3